add language switch option

+++ frontend/src/contexts/CartContext.jsx
@@ -10,6 +10,7 @@ export function CartProvider({ children }) {
   const { user } = useContext(AuthContext);
   const [cart, setCart] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
+  const [isSubmitting, setIsSubmitting] = useState(false);
 
   useEffect(() => {
     const getCart = async () => {
@@ -29,6 +30,8 @@ export function CartProvider({ children }) {
   }, [user?.username]);
 
   const addToCart = async (productId, quantity) => {
+    if (isSubmitting) return;
+    setIsSubmitting(true);
     try {
       const newCartItem = await shoppingCartService.addCartItem(productId, quantity);
       const { id, name, pictureUrl, price } = newCartItem.product;
@@ -40,7 +43,7 @@ export function CartProvider({ children }) {
           return cartItems.map((item) =>
             item.productId === productId ? { ...item, quantity: item.quantity + quantity } : item,
           );
-
+        setIsSubmitting(false);
         return [...cartItems, { id, name, pictureUrl, price, productId, quantity }];
       });
       toast.success('Product(s) added to cart succesfully.');
@@ -50,12 +53,15 @@ export function CartProvider({ children }) {
   };
 
   const removeFromCart = async (productId) => {
+    if (isSubmitting) return;
+    setIsSubmitting(true);
     try {
       await shoppingCartService.removeCartItem(productId);
 
       const newCart = await shoppingCartService.getShoppingCartByUserId();
 
       setCart(newCart);
+      setIsSubmitting(false);
       toast.success('Product(s) deleted from cart succesfully.');
     } catch (error) {
       toast.error('Could not delete product from cart.');
